+++ Pipfile
@@ -15,6 +15,7 @@ bandit = "~=1.5.1"
 isort = "*"
 flake8 = "*"
 autoflake = "*"
+pytest-cov = "*"
 
 [packages]
 django = "~=2.2"
