+++ setup.cfg
@@ -10,3 +10,28 @@ known_first_party=polls,mysite
 line_length=88
 multi_line_output=3
 use_parentheses=True
+
+[tool:pytest]
+DJANGO_SETTINGS_MODULE=mysite.settings
+
+[coverage:run]
+branch = True
+source =
+    polls
+
+[coverage:report]
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain about missing debug-only code:
+    def __repr__
+    if self\.debug
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if 0:
+    if __name__ == .__main__.:
